+++ internal/upgrade/upgrade.go
@@ -18,10 +18,10 @@ import (
 	"github.com/giantswarm/clustertest/v3/pkg/logger"
 	"github.com/giantswarm/clustertest/v3/pkg/wait"
 
-	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/helper"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/state"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/timeout"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/helper"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/state"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/timeout"
 
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 	. "github.com/onsi/gomega"    //nolint:staticcheck