+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactored `common.RunApps` to accept the full `TestConfig` struct instead of individual boolean parameters for better extensibility.
+- Use the `giantswarm-garm-wc-e2e` garm for WC e2e testing on `garm`
 
 ## [3.1.0] - 2026-01-22